# specs/openapi/MULO_OpenAPI_v2.7.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MULO API",
    "version": "2.7",
    "description": "MULO 통합 API 계약 기준 OpenAPI 문서입니다.\n\n- 현재 feature 구현과 기존 OpenAPI v2.6 계약을 함께 보존합니다.\n- 각 작업의 x-implementation-status는 implemented, contract-only, excluded-v1 중 하나입니다.\n- 홈 인기 자물쇠는 조회 시점 기준 최근 7일 데이터를 사용합니다.\n- 내 자물쇠 조회와 사용자 자물쇠 지역 그룹에는 7일 제한을 적용하지 않습니다.\n- 법정동 조회가 정상 완료됐지만 결과가 없으면 legalDongCode/legalDongName을 null로 허용합니다.\n- 커서 기반 목록 API size는 기본 20, 최소 1, 최대 100이며 채팅 메시지 내역만 기본 30입니다.\n- Refresh/CSRF Cookie의 SameSite 최종 정책은 배포 구조에 따라 확정하며 현재 구현은 Strict입니다.\n- 서버 URL은 아직 확정되지 않아 지정하지 않았습니다."
  },
...
  "paths": {
...
    "/api/places/popular": {
      "get": {
        "tags": [
          "HOME · PLACE / 홈 · 장소"
        ],
        "summary": "인기 장소 마커 조회",
        "operationId": "getApiPlacesPopular",
        "description": "홈 인기 자물쇠 지도용 장소 마커를 조회한다.\n\n- viewport는 `swLat`, `swLng`, `neLat`, `neLng`로 전달한다.\n- 조회 시점 기준 최근 7일 이내 생성되고 삭제되지 않은 자물쇠만 대상이다.\n- 최근 7일 내 자물쇠가 하나 이상 존재하는 장소만 반환한다.\n- 각 장소는 `placeId`, `legalDongCode`, `legalDongName`, 좌표, `recordsCount`를 반환한다.\n- `recordsCount`는 해당 장소의 최근 7일 자물쇠 수다.\n- 지도 클러스터링은 프론트에서 처리하며 클러스터 선택 시 포함된 `placeId`들을 다중 장소 인기 음악 API에 전달한다.\n- 비로그인 사용자도 조회할 수 있다.",
        "responses": {
          "200": {
            "description": "현재 지도 영역 조회가 정상 수행되었으므로 200.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "placeId": {
                            "type": "integer",
                            "format": "int64"
                          },
                          "legalDongName": {
                            "type": "string",
                            "nullable": true
                          },
                          "latitude": {
                            "type": "number",
                            "format": "double"
                          },
                          "longitude": {
                            "type": "number",
                            "format": "double"
                          },
                          "legalDongCode": {
                            "type": "string",
                            "nullable": true,
                            "example": "4159012700"
                          },
                          "recordsCount": {
                            "type": "integer",
                            "format": "int64",
                            "minimum": 0
                          }
                        }
                      }
                    }
                  }
                },
                "example": {
                  "message": "인기 장소 조회 성공",
                  "data": [
                    {
                      "placeId": 1,
                      "legalDongName": "오산동",
                      "latitude": 37.2002,
                      "longitude": 127.095,
                      "legalDongCode": "4159012700",
                      "recordsCount": 1
                    },
                    {
                      "placeId": 2,
                      "legalDongName": "반송동",
                      "latitude": 37.1987,
                      "longitude": 127.0723,
                      "legalDongCode": "4159012700",
                      "recordsCount": 1
                    },
                    {
                      "placeId": 3,
                      "legalDongName": "병점동",
                      "latitude": 37.1867,
                      "longitude": 127.072,
                      "legalDongCode": "4159012700",
                      "recordsCount": 1
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "지도 범위를 구성하는 데 필요한 필수 Query 값이 누락되어 유효성 검증에 실패했으므로 400.\n\n추가 케이스: 전달된 지도 좌표가 허용 가능한 위도/경도 범위를 벗어나므로 400.\n\n추가 케이스: 개별 좌표 값은 유효하지만 전달된 좌표들로 정상적인 지도 영역을 구성할 수 없으므로 400.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "field": {
                            "type": "string"
                          },
                          "code": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                },
                "example": {
                  "code": "INVALID_INPUT_VALUE",
                  "message": "입력값을 확인해주세요.",
                  "errors": [
                    {
                      "field": "swLat",
                      "code": "SW_LAT_REQUIRED",
                      "message": "남서쪽 위도 값이 필요합니다."
                    },
                    {
                      "field": "swLng",
                      "code": "SW_LNG_REQUIRED",
                      "message": "남서쪽 경도 값이 필요합니다."
                    },
                    {
                      "field": "neLat",
                      "code": "NE_LAT_REQUIRED",
                      "message": "북동쪽 위도 값이 필요합니다."
                    },
                    {
                      "field": "neLng",
                      "code": "NE_LNG_REQUIRED",
                      "message": "북동쪽 경도 값이 필요합니다."
                    }
                  ]
                }
              }
            }
          },
          "500": {
            "description": "MULO 백엔드 내부의 예상하지 못한 처리 실패이므로 500.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "code": "INTERNAL_SERVER_ERROR",
                  "message": "서버 내부 오류가 발생했습니다. 잠시 후 다시 시도해주세요."
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "swLat",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "swLng",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "neLat",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "neLng",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "format": "double"
            }
          }
        ],
        "x-implementation-status": "contract-only"
      }
    },
...
    "/api/places/popular-tracks/search": {
      "post": {
        "tags": [
          "HOME · PLACE / 홈 · 장소"
        ],
        "summary": "클러스터/장소 인기 음악 조회",
        "description": "필수\nplaceIds: List<Long> (1개 이상)\n\n선택\n없음\n\n정책\n- 비로그인 사용자도 조회할 수 있다.\n- 전달받은 모든 placeIds를 하나의 조회 범위로 합산한다.\n- Record의 placeId가 placeIds에 포함되는 기록만 집계한다.\n- 삭제된 Record는 제외한다.\n- API 조회 시점을 기준으로\n직전 7일 동안 생성된 Record만 집계한다.\n- recordCount는 선택된 placeIds 전체의\n최근 7일 활성 Record 총개수이다.\n- music[].count는 최근 7일 활성 Record 중\n해당 음악이 등록된 횟수이다.\n- 음악별 count DESC로 정렬한다.\n- count가 같으면 해당 음악의\n가장 최근 Record 생성 시각이 최신인 음악을 우선한다.\n- 클러스터링은 프론트의\nKakao MarkerClusterer가 담당한다.\n- 백엔드는 클러스터를 관리하지 않고\n전달받은 placeIds만 처리한다.\n- 개별 마커 클릭은 placeIds에\n1개의 ID를 전달해 동일하게 처리한다.\n- Cursor pagination과 클라이언트 size 입력은\n사용하지 않는다.\n\n정책 및 설명\n- 선택된 placeIds 전체의 삭제되지 않은 Record를 합산한다.\n- 조회 시점 기준 직전 7일의 Record만 집계한다.\n- recordCount와 music[].count 모두 최근 7일 기준이다.\n- 음악별 count DESC로 정렬하고 동률이면\n가장 최근 Record 생성 시각이 최신인 음악을 우선한다.\n- 개별 마커와 클러스터 클릭을 같은 API로 처리한다.\n- 클러스터 ID나 행정구역 정보는 전달하지 않는다.\n- Cursor pagination, size, TOP N 제한은 현재 명세에 적용하지 않는다.",
        "operationId": "post_api_places_popular_tracks_search",
        "security": [

        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "placeIds": {
                    "type": "array",
                    "items": {
                      "type": "integer",
                      "format": "int64",
                      "minimum": 1
                    }
                  }
                },
                "required": [
                  "placeIds"
                ]
              },
              "example": {
                "placeIds": [
                  10,
                  20,
                  30
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "선택된 Place 전체의 최근 7일 인기 음악 집계를 정상 조회했으므로 200.\n\n최근 7일 활성 Record가 없어도 정상 조회이므로 200과 빈 배열을 반환한다.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "recordCount": {
                          "type": "integer"
                        },
                        "music": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "rank": {
                                "type": "integer"
                              },
                              "musicTrackId": {
                                "type": "integer",
                                "format": "int64"
                              },
                              "title": {
                                "type": "string"
                              },
                              "artistName": {
                                "type": "string"
                              },
                              "count": {
                                "type": "integer"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                },
                "examples": {
                  "example1": {
                    "value": {
                      "message": "인기 음악 조회 성공",
                      "data": {
                        "recordCount": 214,
                        "music": [
                          {
                            "rank": 1,
                            "musicTrackId": 583,
                            "title": "밤편지",
                            "artistName": "아이유",
                            "count": 42
                          },
                          {
                            "rank": 2,
                            "musicTrackId": 921,
                            "title": "Ditto",
                            "artistName": "NewJeans",
                            "count": 37
                          }
                        ]
                      }
                    }
                  },
                  "example2": {
                    "value": {
                      "message": "인기 음악 조회 성공",
                      "data": {
                        "recordCount": 0,
                        "music": [

                        ]
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "placeIds가 누락되었거나 빈 배열이므로 400.\n\nplaceIds에 유효하지 않은 장소 ID 형식이 포함되어 있으므로 400.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "field": {
                            "type": "string"
                          },
                          "code": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                },
                "examples": {
                  "example1": {
                    "value": {
                      "code": "INVALID_INPUT_VALUE",
                      "message": "입력값을 확인해주세요.",
                      "errors": [
                        {
                          "field": "placeIds",
                          "code": "PLACE_IDS_REQUIRED",
                          "message": "하나 이상의 장소 ID가 필요합니다."
                        }
                      ]
                    }
                  },
                  "example2": {
                    "value": {
                      "code": "INVALID_INPUT_VALUE",
                      "message": "입력값을 확인해주세요.",
                      "errors": [
                        {
                          "field": "placeIds",
                          "code": "INVALID_PLACE_ID",
                          "message": "올바른 장소 ID를 입력해주세요."
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "MULO 백엔드 내부의 예상하지 못한 처리 실패이므로 500.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "examples": {
                  "example1": {
                    "value": {
                      "code": "INTERNAL_SERVER_ERROR",
                      "message": "서버 내부 오류가 발생했습니다. 잠시 후 다시 시도해주세요."
                    }
                  }
                }
              }
            }
          }
        },
        "x-implementation-status": "contract-only",
        "x-csrf-required": true
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Access Token 인증이 필요한 API에서 사용"
      },
      "refreshCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refreshToken",
        "description": "HttpOnly Refresh Token Cookie. Path=/api/auth, 운영 Secure=true. SameSite 최종 정책은 배포 구조에 따라 확정하며 현재 구현은 Strict."
      },
      "xsrfHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-XSRF-TOKEN",
        "description": "GET /api/csrf가 발급한 XSRF-TOKEN Cookie의 원본 값을 상태 변경 요청에 전달한다. POST /api/users/signup은 예외다."
      }
    },
...
  }
}